[WIP] udn bgp support
Create an external aws instance on the same subnet as worker node.
Run external frr on this external aws instance.
Run kube burner workload on this external aws instance.

+++ workloads/kube-burner-ocp-wrapper/egressip.sh
@@ -7,11 +7,9 @@ export AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY:-""}
 
 TEMP_DIR=`mktemp -d`
 
-prep_aws(){
+install_oc_cli(){
     if [[ -z $(go version) ]]; then
-        curl -L https://go.dev/dl/go1.18.2.linux-amd64.tar.gz -o go1.18.2.linux-amd64.tar.gz
-        tar -C ${TEMP_DIR}/ -xzf go1.18.2.linux-amd64.tar.gz
-        export PATH=${TEMP_DIR}/go/bin:$PATH
+        dnf install golang -y
     fi
     if [[ -z $(rosa version)  ]]; then
         mkdir -p ${TEMP_DIR}/bin/
@@ -25,20 +23,31 @@ prep_aws(){
         tar xzvf openshift-client-linux.tar.gz
         mv oc kubectl ${TEMP_DIR}/bin/
     fi
+}
+
+install_aws(){
     if [[ -z $(aws --version) ]]; then
         curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
         unzip awscliv2.zip
         ./aws/install --install-dir "$TEMP_DIR/.aws-cli" --bin-dir "$TEMP_DIR/.aws-cli/bin" --update
         export PATH="$TEMP_DIR/.aws-cli/bin:$PATH"
     fi
-    echo [default] > ${TEMP_DIR}/aws_credentials
-    echo aws_access_key_id=$AWS_ACCESS_KEY_ID >> ${TEMP_DIR}/aws_credentials
-    echo aws_secret_access_key=$AWS_SECRET_ACCESS_KEY >> ${TEMP_DIR}/aws_credentials
-    echo [default] > ${TEMP_DIR}/aws_config
-    echo region=$AWS_REGION >> ${TEMP_DIR}/aws_config
-    echo output=json >> ${TEMP_DIR}/aws_config
-    export AWS_SHARED_CREDENTIALS_FILE=${AWS_SHARED_CREDENTIALS_FILE:-${TEMP_DIR}/aws_credentials}
-    export AWS_CONFIG_FILE=${TEMP_DIR}/aws_config
+	if [[ -z "${AWS_ACCESS_KEY_ID}" || -z "${AWS_SECRET_ACCESS_KEY}" ]]; then
+		export AWS_SHARED_CREDENTIALS_FILE="${CLUSTER_PROFILE_DIR}/.awscred"
+		if [[ ! -f $AWS_SHARED_CREDENTIALS_FILE ]]; then
+			echo "ERROR: AWS credentials file ${AWS_SHARED_CREDENTIALS_FILE} does not exist."
+			exit 1
+		fi
+	else
+        echo [default] > ${TEMP_DIR}/aws_credentials
+        echo aws_access_key_id=$AWS_ACCESS_KEY_ID >> ${TEMP_DIR}/aws_credentials
+        echo aws_secret_access_key=$AWS_SECRET_ACCESS_KEY >> ${TEMP_DIR}/aws_credentials
+        echo [default] > ${TEMP_DIR}/aws_config
+        echo region=$AWS_REGION >> ${TEMP_DIR}/aws_config
+        echo output=json >> ${TEMP_DIR}/aws_config
+        export AWS_SHARED_CREDENTIALS_FILE=${AWS_SHARED_CREDENTIALS_FILE:-${TEMP_DIR}/aws_credentials}
+        export AWS_CONFIG_FILE=${TEMP_DIR}/aws_config
+	fi
 }
 
 create_egressip_external_server(){
@@ -52,13 +61,77 @@ create_egressip_external_server(){
     USER_DATA_SCRIPT="user-data.sh"
     REGION="us-west-2"
 
+    for sg_id in $SECURITY_GROUP_IDS; do
+        name=$(aws ec2 describe-security-groups --group-ids "$sg_id" --query 'SecurityGroups[0].GroupName' --output text)
+        if [[ "$name" == *-node ]]; then
+		    aws ec2 authorize-security-group-ingress --group-id $sg_id --protocol tcp --port 179 --source-group $sg_id
+            break
+        fi
+    done
+
+# Check if KUBECONFIG is set
+if [ -z "$KUBECONFIG" ]; then
+  echo "KUBECONFIG environment variable is not set."
+  exit 1
+fi
+
+# Check if the file exists
+if [ ! -f "$KUBECONFIG" ]; then
+  echo "Kubeconfig file not found: $KUBECONFIG"
+  exit 1
+fi
+# Read the content into a variable
+KUBECONFIG_CONTENT=$(< "$KUBECONFIG")
+
+CONFIG_VTYSH_CMDS=$(cat <<'VTYSH_CMDS_DELIMITER'
+configure terminal
+router bgp 64512
+redistribute static
+redistribute connected
+end
+write
+VTYSH_CMDS_DELIMITER
+)
+
     # Create the user-data script
     cat <<EOF > $USER_DATA_SCRIPT
 #!/bin/bash
-sudo dnf install podman -y
-for port in {9002..9020}; do
-    podman run --network=host -d -e LISTEN_PORT=\$port quay.io/cloud-bulldozer/nginxecho:latest
-done
+set -x
+
+echo "root:temppass" | sudo chpasswd
+sudo sed -i 's/^PasswordAuthentication no/PasswordAuthentication yes/' /etc/ssh/sshd_config
+sudo systemctl restart sshd
+
+sudo dnf install curl git make binutils bison gcc glibc-devel golang tmux podman jq -y
+mkdir -p /root/.kube/
+
+# Write the kubeconfig file
+cat <<KCFG > /root/.kube/config
+$KUBECONFIG_CONTENT
+KCFG
+
+chmod 666 /root/.kube/config
+export KUBECONFIG=/root/.kube/config
+
+cd /tmp
+curl -sSL https://mirror.openshift.com/pub/openshift-v4/clients/ocp/stable/openshift-client-linux-amd64-rhel8.tar.gz | tar -xvzf -
+mv oc /usr/local/bin/
+mv kubectl /usr/local/bin/
+
+export GOCACHE=/tmp/gocache
+mkdir -p $GOCACHE
+cd /tmp
+git clone -b ovnk-bgp https://github.com/jcaamano/frr-k8s
+cd frr-k8s/hack/demo/
+./demo.sh
+sleep 30
+oc apply -f configs/receive_all.yaml
+
+echo "${CONFIG_VTYSH_CMDS}" | sudo podman exec -i frr vtysh
+
+curl --fail --retry 8 --retry-all-errors -sS -L "${KUBE_BURNER_URL}" | tar -xzC "${KUBE_DIR}/" kube-burner-ocp
+$COMMAND
+
 EOF
 
     # Note: We use the same name for both key and instance
@@ -108,8 +181,13 @@ fi
 }
 
 get_egressip_external_server(){
+	install_aws
+    oc patch Network.operator.openshift.io cluster --type=merge -p='{"spec":{"additionalRoutingCapabilities": {"providers": ["FRR"]}, "defaultNetwork":{"ovnKubernetesConfig":{"routeAdvertisements":"Enabled"}}}}'
+	# retain this sleep
+    #sleep 300
     CLUSTER_NAME=$(oc get infrastructure cluster -o jsonpath='{.status.infrastructureName}')
-    export EGRESSIP_EXTERNAL_SERVER_INSTANCE_NAME="$CLUSTER_NAME-egress-server"
+	# TODO: revert name back
+    export EGRESSIP_EXTERNAL_SERVER_INSTANCE_NAME="$CLUSTER_NAME-egress-server2"
     EGRESSIP_EXTERNAL_SERVER_INSTANCE_ID=$(aws ec2 describe-instances --filters "Name=tag:Name,Values=${EGRESSIP_EXTERNAL_SERVER_INSTANCE_NAME}" "Name=instance-state-name,Values=running" --query "Reservations[*].Instances[*].InstanceId" --output text)
     # Check if an instance ID was found
     if [ -z "$EGRESSIP_EXTERNAL_SERVER_INSTANCE_ID" ]; then
